Expose clusterIPs so we can easily create a multi-master setup
Calico will announce all the kubernetes services through BGP; this means
we have a stable IP address to reach the apiserver. We can use this to
bring up multiple masters; which are all serving the same apiserver
ClusterIP

This is what makes IPV6 _extremely_ cool and should be the crux of the
story we're telling here. We do not need an external loadbalancer for
multi-master kubernetes;  because all IP addresses are public.

In the IPv6 world your world kinda turns upside down and things become a
lot easier!

However!! note that this means all pods and services are publically
reachable by default! we can't rely on NAT for security anymore (which
you shouldn't do in the first place). Instead; we can use Kubernetes'
NetworkPolicy to  firewall pods and services. This is something we'll
explore next.